requirements.txt
python-novaclient>=2.17
Babel==1.3
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
MySQL-python==1.2.5
PyMySQL==0.6.1
PyYAML==3.10
SQLAlchemy==0.8.0
WebOb==1.3.1
WebTest==2.0.14
Werkzeug==0.9.4
argparse==1.2.1
beautifulsoup4==4.3.2
decorator==3.4.0
httplib2==0.8
iso8601==0.1.8
itsdangerous==0.23
mock==1.0.1
netaddr==0.7.10
nose==1.3.0
oslo.config==1.2.1
pbr==0.6
prettytable==0.7.2
psycopg2==2.5.2
pyaml==13.07.0
python-ceilometerclient==1.0.3
python-keystoneclient==0.3.2
python-novaclient==2.17.0
pytz==2013.9
requests==1.1.0
requirements-parser==0.0.6
simplejson==3.3.3
six==1.5.2
urllib3==1.5
waitress==0.8.8
wsgiref==0.1.2
